@@ -1,13 +1,13 @@
 {
   "name": "react-app-simple-chat-app",
-  "version": "1.0.0",
+  "version": "2.0.0",
   "description": "This is a simple chat Application",
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "build": "webpack --mode production",
-    "dev": "webpack --mode development"
+    "server": "node -r dotenv/config server/server.js dotenv_config_path=server/.env"
   },
+  "type": "module",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/anijitsahu/react-app-simple-chat-app.git"
@@ -20,29 +20,26 @@
   },
   "homepage": "https://github.com/anijitsahu/react-app-simple-chat-app#readme",
   "devDependencies": {
-    "@babel/core": "^7.1.2",
-    "@babel/preset-env": "^7.1.0",
-    "@babel/preset-react": "^7.0.0",
-    "axios": "^0.18.0",
-    "babel-loader": "^8.0.4",
-    "css-loader": "^1.0.0",
-    "file-loader": "^2.0.0",
-    "mini-css-extract-plugin": "^0.4.4",
-    "prop-types": "^15.6.2",
-    "react": "^16.5.2",
-    "react-dom": "^16.5.2",
-    "socket.io-client": "^2.1.1",
-    "style-loader": "^0.23.1",
-    "uuid": "^3.3.2",
-    "webpack": "^4.20.2",
-    "webpack-cli": "^3.1.2"
+    "@babel/core": "^7.13.10",
+    "@babel/preset-react": "^7.12.13",
+    "axios": "^0.21.1",
+    "babel-loader": "^8.2.2",
+    "css-loader": "^5.1.3",
+    "react": "^17.0.1",
+    "react-dom": "^17.0.1",
+    "react-router-dom": "^5.2.0",
+    "socket.io-client": "^4.0.0",
+    "style-loader": "^2.0.0",
+    "webpack": "^5.26.3",
+    "webpack-cli": "^4.5.0"
   },
   "dependencies": {
-    "body-parser": "^1.18.3",
-    "cors": "^2.8.4",
-    "express": "^4.16.4",
-    "mongodb": "^3.1.10",
-    "morgan": "^1.9.1",
-    "socket.io": "^2.1.1"
+    "cors": "^2.8.5",
+    "dotenv": "^8.2.0",
+    "express": "^4.17.1",
+    "helmet": "^4.4.1",
+    "mongodb": "^3.6.5",
+    "morgan": "^1.10.0",
+    "socket.io": "^4.0.0"
   }
 }
